requirements.txt
Django>= 3.0.0,<3.2.1
djangorestframework>=3.0.0
psycopg2>=2.7.5,<2.8.0
pillow>=5.3.0,<5.4.0

mock==4.0.*
factory-boy==3.2.*

celery==5.*
redis==3.5.*
selenium==3.*
django-redis==5.0.*

django-debug-toolbar==3.2.*

flake8>=3.6.0,<=3.7.0